pyproject.toml
[project]
name = "right-of-way-rl"
version = "0.1.0"
description = "RL"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch==2.6.0",
    "torchvision==0.21.0",
    "torchaudio==2.6.0",
    "pandas>=2.2.3",
    "pyzmq>=27.0.0",
    "metadrive-simulator",
    "hydra-core>=1.3.2",
    "tensorboard>=2.19.0",
    "gymnasium>=1.2.0",
]


# 默认走清华 pypi
[[tool.uv.index]]
name = "pypi"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
explicit = false

[[tool.uv.index]]
name = "pytorch-cu126"
url = " https://download.pytorch.org/whl/cu126"
explicit = true

[tool.uv.sources]
metadrive-simulator = { git = "https://github.com/metadriverse/metadrive.git", rev = "main" }
